pyproject.toml
[project]
name = "hk-task-scheduler-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=1.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
